fix revisi

+++ app/Http/Controllers/UserController.php
@@ -64,12 +64,13 @@ class UserController extends Controller
         $user->user_status = 'pengguna';
         if($request->hasFile('photo_identitas'))
         {
-            $destination = public_path().'/user_photo/'.$user->photo_identitas;;
+
             // dd($destination);
-            if(File::exists($destination))
-            {
-                unlink($destination);
+            if(file_exists($user->image)){
+                $image_path = public_path().'/user_photo/'.$user->image;
+                unlink($image_path);
             }
+
             $file = $request->file('photo_identitas');
             $extention = $file->getClientOriginalExtension();
             $filename=time().'.'.$extention;
@@ -98,8 +99,12 @@ class UserController extends Controller
 
     public function delete($id){
         $user = User::find($id);
-        $image_path = public_path().'/user_photo/'.$user->photo_identitas;
-        unlink($image_path);
+        // dd($destination);
+        if(file_exists($user->photo_identitas)){
+            $image_path = public_path().'/user_photo/'.$user->image;
+            unlink($image_path);
+        }
+
         $user->delete();
         return redirect('/user');
     }
